fix
E   pydantic.errors.PydanticUserError: A non-annotated attribute was detected: `max_process_count = 10`. All model fields require a type annotation; if `max_process_count` is not meant to be a field, you may be able to resolve this error by annotating it as a `ClassVar` or updating `model_config['ignored_types']`.